+++ org.dieschnittstelle.jee.esa.jsf.cdi/src/org/dieschnittstelle/jee/esa/jsf/controller/ProductsViewController.java
@@ -89,7 +89,7 @@ public class ProductsViewController implements Serializable {
 		// add the product to the cart
 		logProductBundleKlass();
 		
-		AbstractProduct product = findProduct(Integer.parseInt(id));
+		AbstractProduct product = findProduct(Long.parseLong(id));
 		CrmProductBundle productBundle = new CrmProductBundle(product.getId(),
 				1);
 		productBundle.setProductObj(product);
@@ -126,7 +126,7 @@ public class ProductsViewController implements Serializable {
 	 * @param id
 	 * @return
 	 */
-	private AbstractProduct findProduct(int id) {
+	private AbstractProduct findProduct(long id) {
 		for (AbstractProduct prod : this.products) {
 			if (prod.getId() == id)
 				return prod;
